+++ src/byte/lib/types/__init__.py
@@ -1,0 +1,5 @@
+"""Types and similar facilities used throughout library code."""
+
+from byte.lib.types import astral, python
+
+__all__ = ("astral", "python")
