+++ src/byte/lib/utils.py
@@ -1,4 +1,5 @@
 """Byte utilities."""
+
 from __future__ import annotations
 
 import datetime as dt
@@ -6,197 +7,34 @@ import json
 import re
 import subprocess
 from datetime import UTC, datetime
-from enum import StrEnum
 from itertools import islice
-from typing import TYPE_CHECKING, TypedDict, TypeVar
+from typing import TYPE_CHECKING, TypeVar
 
 import httpx
 from anyio import run_process
-from discord.ext import commands
-from discord.ext.commands import CheckFailure
 from ruff.__main__ import find_ruff_bin  # type: ignore[import-untyped]
 
-from byte.lib import settings
 from byte.lib.common.links import pastebin
+from byte.lib.types.python import PEP, PEPStatus, PEPType
 
 if TYPE_CHECKING:
     from collections.abc import Iterable
-    from typing import Any
 
-    from discord.ext.commands import Context
-    from discord.ext.commands._types import Check
+    from byte.lib.types.astral import FormattedRuffRule, RuffRule
 
 __all__ = (
-    "BaseRuffRule",
-    "RuffRule",
-    "FormattedRuffRule",
-    "PEP",
-    "PEPType",
-    "PEPStatus",
-    "PEPHistoryItem",
-    "is_guild_admin",
-    "is_byte_dev",
-    "linker",
-    "mention_user",
-    "mention_user_nickname",
-    "mention_channel",
-    "mention_role",
-    "mention_slash_command",
-    "mention_custom_emoji",
-    "mention_custom_emoji_animated",
-    "mention_timestamp",
-    "mention_guild_navigation",
+    "chunk_sequence",
+    "format_resolution_link",
     "format_ruff_rule",
+    "get_next_friday",
+    "linker",
+    "paste",
+    "query_all_peps",
     "query_all_ruff_rules",
     "run_ruff_format",
-    "paste",
-    "chunk_sequence",
-    "query_all_peps",
-    "get_next_friday",
 )
 
 _T = TypeVar("_T")
-
-
-class BaseRuffRule(TypedDict):
-    """Base Ruff rule data."""
-
-    name: str
-    summary: str
-    fix: str
-    explanation: str
-
-
-class RuffRule(BaseRuffRule):
-    """Ruff rule data."""
-
-    code: str
-    linter: str
-    message_formats: list[str]
-    preview: bool
-
-
-class FormattedRuffRule(BaseRuffRule):
-    """Formatted Ruff rule data."""
-
-    rule_link: str
-    rule_anchor_link: str
-
-
-class PEPType(StrEnum):
-    """Type of PEP.
-
-    Based off of `PEP Types in PEP1 <https://peps.python.org/#pep-types-key>`_.
-    """
-
-    I = "Informational"  # noqa: E741
-    P = "Process"
-    S = "Standards Track"
-
-
-class PEPStatus(StrEnum):
-    """Status of a PEP.
-
-    .. note:: ``Active`` and ``Accepted`` both traditionally use ``A``,
-        but are differentiated here for clarity.
-
-    Based off of `PEP Status in PEP1 <https://peps.python.org/#pep-status-key>`_.
-    """
-
-    A = "Active"
-    AA = "Accepted"
-    D = "Deferred"
-    __ = "Draft"
-    F = "Final"
-    P = "Provisional"
-    R = "Rejected"
-    S = "Superseded"
-    W = "Withdrawn"
-
-
-class PEPHistoryItem(TypedDict, total=False):
-    """PEP history item.
-
-    Sometimes these include a list of ``datetime`` objects,
-    other times they are a list of datetime and str
-    because they contain a date and an rST link.
-    """
-
-    date: str
-    link: str
-
-
-class PEP(TypedDict):
-    """PEP data.
-
-    Based off of the `PEPS API <https://peps.python.org/api/peps.json>`_.
-    """
-
-    number: int
-    title: str
-    authors: list[str] | str
-    discussions_to: str
-    status: PEPStatus
-    type: PEPType
-    topic: str
-    created: datetime
-    python_version: list[float] | float
-    post_history: list[str]
-    resolution: str | None
-    requires: str | None
-    replaces: str | None
-    superseded_by: str | None
-    url: str
-
-
-def is_guild_admin() -> Check[Any]:
-    """Check if the user is a guild admin.
-
-    Returns:
-        A check function.
-    """
-
-    async def predicate(ctx: Context) -> bool:
-        """Check if the user is a guild admin.
-
-        Args:
-            ctx: Context object.
-
-        Returns:
-            True if the user is a guild admin, False otherwise.
-        """
-        member = ctx.guild.get_member(ctx.author.id)
-        if not member:
-            msg = "Member not found in the guild."
-            raise CheckFailure(msg)
-
-        return member.guild_permissions.administrator
-
-    return commands.check(predicate)
-
-
-def is_byte_dev() -> Check[Any]:
-    """Check if the user is a Byte developer.
-
-    Returns:
-        A check function.
-    """
-
-    async def predicate(ctx: Context) -> bool:
-        """Check if the user is a Byte Dev or Owner.
-
-        Args:
-            ctx: Context object.
-
-        Returns:
-            True if the user is a Byte Dev or Owner, False otherwise.
-        """
-        if await ctx.bot.is_owner(ctx.author) or ctx.author.id == settings.discord.DEV_USER_ID:
-            return True
-
-        return any(role.name == "byte-dev" for role in ctx.author.roles)  # type: ignore[reportAttributeAccessIssue]
-
-    return commands.check(predicate)
 
 
 def linker(title: str, link: str, show_embed: bool = False) -> str:
@@ -211,120 +49,6 @@ def linker(title: str, link: str, show_embed: bool = False) -> str:
         A Markdown link.
     """
     return f"[{title}]({link})" if show_embed else f"[{title}](<{link}>)"
-
-
-def mention_user(user_id: int) -> str:
-    """Mention a user by ID.
-
-    Args:
-        user_id: The unique identifier for the user.
-
-    Returns:
-        A formatted string that mentions the user.
-    """
-    return f"<@{user_id}>"
-
-
-def mention_user_nickname(user_id: int) -> str:
-    """Mention a user by ID with a nickname.
-
-    Args:
-        user_id: The unique identifier for the user.
-
-    Returns:
-        A formatted string that mentions the user with a nickname.
-    """
-    return f"<@!{user_id}>"
-
-
-def mention_channel(channel_id: int) -> str:
-    """Mention a channel by ID.
-
-    Args:
-        channel_id: The unique identifier for the channel.
-
-    Returns:
-        A formatted string that mentions the channel.
-    """
-    return f"<#{channel_id}>"
-
-
-def mention_role(role_id: int) -> str:
-    """Mention a role by ID.
-
-    Args:
-        role_id: The unique identifier for the role.
-
-    Returns:
-        A formatted string that mentions the role.
-    """
-    return f"<@&{role_id}>"
-
-
-def mention_slash_command(name: str, command_id: int) -> str:
-    """Mention a slash command by name and ID.
-
-    Args:
-        name: The name of the slash command.
-        command_id: The unique identifier for the slash command.
-
-    Returns:
-        A formatted string that mentions the slash command.
-    """
-    return f"</{name}:{command_id}>"
-
-
-def mention_custom_emoji(name: str, emoji_id: int) -> str:
-    """Mention a custom emoji by name and ID.
-
-    Args:
-        name: The name of the emoji.
-        emoji_id: The unique identifier for the emoji.
-
-    Returns:
-        A formatted string that mentions the custom emoji.
-    """
-    return f"<:{name}:{emoji_id}>"
-
-
-def mention_custom_emoji_animated(name: str, emoji_id: int) -> str:
-    """Mention an animated custom emoji by name and ID.
-
-    Args:
-        name: The name of the animated emoji.
-        emoji_id: The unique identifier for the animated emoji.
-
-    Returns:
-        A formatted string that mentions the animated custom emoji.
-    """
-    return f"<a:{name}:{emoji_id}>"
-
-
-def mention_timestamp(timestamp: int, style: str = "") -> str:
-    """Mention a timestamp, optionally with a style.
-
-    Args:
-        timestamp: The Unix timestamp to format.
-        style: An optional string representing the timestamp style.
-               (Default `` ``, valid styles: ``t``, ``T``, ``d``, ``D``, ``f``, ``F``, ``R``)
-
-    Returns:
-        A formatted string that represents the timestamp.
-    """
-    return f"<t:{timestamp}:{style}>" if style else f"<t:{timestamp}>"
-
-
-def mention_guild_navigation(guild_nav_type: str, guild_element_id: int) -> str:
-    """Mention a guild navigation element by type and ID.
-
-    Args:
-        guild_nav_type: The type of the guild navigation element.
-        guild_element_id: The unique identifier for the element.
-
-    Returns:
-        A formatted string that mentions the guild navigation element.
-    """
-    return f"<{guild_element_id}:{guild_nav_type}>"
 
 
 def format_ruff_rule(rule_data: RuffRule) -> FormattedRuffRule:
